20 fetch advert detail

+++ MyEstate/app/src/main/java/com/example/myestate/ui/screens/home/HomeScreen.kt
@@ -21,7 +21,7 @@ import org.w3c.dom.Text
 
 @Composable
 fun HomeScreen(
-    navigateToDetail: () -> Unit,
+    navigateToDetail: (Int) -> Unit,
     viewModel: HomeViewModel = hiltViewModel()
 ) {
     val uiState by viewModel.state.collectAsState()
